feat: Add launcher activity and rename other views
