+++ Vinylsg12/app/src/main/AndroidManifest.xml
@@ -15,22 +15,27 @@
         android:theme="@style/Theme.VinylsG12"
         tools:targetApi="31">
         <activity
-            android:name=".ui.about_us.AboutActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.album.AlbumActivity"
-            android:exported="false" />
-        <activity
             android:name=".MainActivity"
-            android:exported="true"
-            android:label="@string/app_name"
-            android:theme="@style/Theme.AppCompat.NoActionBar">
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".ui.about_us.AboutActivity"
+            android:exported="false" />
+        <activity
+            android:name=".ui.album.AlbumActivity"
+            android:exported="false" />
+        <activity
+            android:name=".HomeActivity"
+            android:exported="true"
+            android:label="@string/app_name"
+            android:theme="@style/Theme.AppCompat.NoActionBar"
+            android:parentActivityName=".MainActivity"
+            />
     </application>
 
 </manifest>
