+++ Vinylsg12/app/src/main/java/co/edu/uniandes/vinylsg12/MainActivity.kt
@@ -1,95 +1,11 @@
 package co.edu.uniandes.vinylsg12
 
-import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
-import android.view.Window
-import androidx.appcompat.app.ActionBarDrawerToggle
-import com.google.android.material.bottomnavigation.BottomNavigationView
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.GravityCompat
-import androidx.drawerlayout.widget.DrawerLayout
-import androidx.navigation.NavController
-import androidx.navigation.findNavController
-import androidx.navigation.ui.AppBarConfiguration
-import androidx.navigation.ui.setupActionBarWithNavController
-import androidx.navigation.ui.setupWithNavController
-import co.edu.uniandes.vinylsg12.databinding.ActivityMainBinding
-import co.edu.uniandes.vinylsg12.ui.about_us.AboutActivity
-import com.google.android.material.navigation.NavigationView
+import android.os.Bundle
 
-class MainActivity : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListener {
-
-    private lateinit var binding: ActivityMainBinding
-    private lateinit var drawerLayout: DrawerLayout
-    private lateinit var toggle: ActionBarDrawerToggle
-    private val destinationChangedListener =
-        NavController.OnDestinationChangedListener { _, _, _ ->
-            toggle.syncState()
-        }
-
+class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        setSupportActionBar(binding.toolbar)
-
-        val navView: BottomNavigationView = binding.navView
-        drawerLayout = binding.drawerLayout
-
-        val navMenuView = binding.navMenuView
-
-        val navController = findNavController(R.id.nav_host_fragment_activity_main)
-        // Passing each menu ID as a set of Ids because each
-        // menu should be considered as top level destinations.
-        val appBarConfiguration = AppBarConfiguration(
-            setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_notifications
-            )
-        )
-        setupActionBarWithNavController(navController, appBarConfiguration)
-        navView.setupWithNavController(navController)
-
-        navMenuView.setNavigationItemSelectedListener(this)
-
-       toggle = ActionBarDrawerToggle(
-            this,
-            drawerLayout,
-            binding.toolbar,
-            R.string.navigation_drawer_open,
-            R.string.navigation_drawer_close
-        )
-        drawerLayout.addDrawerListener(toggle)
-        toggle.syncState()
-
-        navController.addOnDestinationChangedListener(destinationChangedListener)
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        val navController = findNavController(R.id.nav_host_fragment_activity_main)
-        navController.removeOnDestinationChangedListener(destinationChangedListener)
-    }
-
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.main_menu, menu)
-        return super.onCreateOptionsMenu(menu)
-    }
-
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        if(item?.itemId == R.id.action_about_us) {
-            AboutActivity.start(context = this)
-            return true
-        }
-        return super.onOptionsItemSelected(item)
-    }
-
-    override fun onNavigationItemSelected(item: MenuItem): Boolean {
-        drawerLayout.closeDrawer(GravityCompat.START)
-        when(item.itemId) {
-            R.id.action_about_us -> AboutActivity.start(this)
-        }
-        return true
+        setContentView(R.layout.activity_main)
     }
 }
