+++ Vinylsg12/app/src/main/java/co/edu/uniandes/vinylsg12/ui/collectors/CollectorsFragment.kt
@@ -1,4 +1,4 @@
-package co.edu.uniandes.vinylsg12.ui.dashboard
+package co.edu.uniandes.vinylsg12.ui.collectors
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import co.edu.uniandes.vinylsg12.databinding.FragmentDashboardBinding
+import co.edu.uniandes.vinylsg12.databinding.FragmentCollectorsBinding
 
-class DashboardFragment : Fragment() {
+class CollectorsFragment : Fragment() {
 
-    private var _binding: FragmentDashboardBinding? = null
+    private var _binding: FragmentCollectorsBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class DashboardFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val dashboardViewModel =
-            ViewModelProvider(this).get(DashboardViewModel::class.java)
+        val collectorsViewModel =
+            ViewModelProvider(this).get(CollectorsViewModel::class.java)
 
-        _binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        _binding = FragmentCollectorsBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
         val textView: TextView = binding.textDashboard
-        dashboardViewModel.text.observe(viewLifecycleOwner) {
+        collectorsViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }
         return root
