+++ Vinylsg12/app/src/main/java/co/edu/uniandes/vinylsg12/ui/collectors/CollectorsViewModel.kt
@@ -1,10 +1,10 @@
-package co.edu.uniandes.vinylsg12.ui.dashboard
+package co.edu.uniandes.vinylsg12.ui.collectors
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class CollectorsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
