+++ Vinylsg12/app/src/main/res/layout/activity_main_nav.xml
@@ -6,7 +6,7 @@
     android:id="@+id/drawer_layout"
     android:fitsSystemWindows="true">
 
-    <include layout="@layout/activity_main"
+    <include layout="@layout/activity_home"
         android:layout_height="match_parent"
         android:layout_width="match_parent"
         />
