+++ Vinylsg12/app/src/main/res/layout/fragment_albums.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.HomeFragment">
+    tools:context=".ui.albums.HomeFragment">
 
     <ImageView
         android:id="@+id/imageView2"
