+++ Vinylsg12/app/src/main/res/values/dimens.xml
@@ -4,4 +4,7 @@
     <dimen name="activity_vertical_margin">16dp</dimen>
     <dimen name="header_vinyls_margin">8dp</dimen>
     <dimen name="header_vinyls_size">30dp</dimen>
+
+    <dimen name="logo_width">113dp</dimen>
+    <dimen name="logo_height">125dp</dimen>
 </resources>
